requirements.txt
aiogram==3.20.0.post0
python-dotenv
pandas
openpyxl
SpeechRecognition
pydub
mistralai
rapidfuzz
nltk